write some insertion sort in c
seems to work. currently only works by passing in arguments in command line. tried to write it so that you can pass an input file of arbitrary size and yowza that's hard. shelved it for now, i'm not really looking to be a c programmer.

played around with a makefile as well, seems useful for bigger stuff, maybe will get more used to it later.